Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > > Patch Set 1:
> > > > > 
> > > > > Andrew, is there any reason we can't just invoke 'modprobe' to load the named module in the chain, and that will do a similar thing and load all of the module's dependencies?
> > > > > 
> > > > 
> > > > This can be done, and is actually how I originally implemented this. The first thing that needed to be done was to fixup the modprobe shipped with toybox, it currently does not search the right Android paths for modules (vendor/). When doing this I realized how little code would be needed if this was all just done in init. We would also then not depend on an external modprobe and init could even become the systems modprobe tool (like with ueventd and watchdogd).
> > > > 
> > > > I'm okay with either if there are strong opinions one way or another.
> > > 
> > > I think we want to go the modprobe route, just because we will always have modprobe, and it already has the logic to parse modules.dep.
> > 
> > I went back and gave that a try, the issue I'm getting with toybox's modprobe is that it has a lot of output that is useful when used as a stand-alone tool, but spams the boot log when used by init for each modalias uevent.
> 
> that's fine. let me know what the output should be and i'll fix toybox modprobe. is the "real" modprobe silent except on error or if you supply -v?
> 


You can make the real modprobe quiet with -q. But with toybox for every event we still get:

modprobe: /etc/modprobe.conf: No such file or directory
modprobe: /etc/modprobe.d: No such file or directory

It also returns an error code when it fails to find a modules for a modalias, when used as a command line tool that makes sense, but when used from init it should not be an error. I'm not sure how to best differentiate these two uses without breaking the command line version.

I have to back-track on my indifference earlier about using an external modprobe or not. With more testing I am now convinced it should be internal to init like is done in this patch.

Otherwise every modalias event has to be handled by starting an instance of modprobe, which then goes and re-reads and re-parses all the needed files, only to find out 99% of the time the modalias event did not require any module to be loaded. There are hundreds of such events on every boot.

Keeping this internal to init allows us to instantly ignore these extra events and only read/parse the needed module information files once, keeping that data in an simple lookup map. For systems without modules (most right now) the module dependency file will be empty or non-existent, in that case we simply return immediately from a modalias event without any extra processing needed.

I can post the external modprobe version with another change ID if that is preferred.

Andrew


> > I'm leaning back now to using this patch as is and having init do the modules.dep parsing, it is quite trivial and reduces an external dependency. We also get some added security doing it ourselves as we read the module and other files with ReadFile and that performs some checks.
> > 
> > > 
> > > > > I can see one disadvantage is that modprobe is fire-and-forget, there's no recourse if it fails, but I'm not sure what error handling you could usefully do from ueventd/init anyway.
> > > > 
> > > > If one attempts to load modules before the filesystem they are on is ready we can do deferred loading, much like how is done with firmware in init/firmware_handler.cpp.
> > > 
> > > Is this a use case that you have? If so, is there a reason those drivers aren't already compiled into the kernel?

Patch-set: 1
Reviewer: Gerrit User 1339568 <1339568@85c56323-6fa9-3386-8a01-6480fb634889>