Update patch set 2
Patch Set 2:

What is the plan with internal vs external lmkd?  Is the internal version with statsd going to be the final version?  In which case is this needed?

How are these APIs used currently?  Is it that you can only ever call android_log_write_list() once per create_android_logger() call and the goal here is to be able to reuse this result?  What is stopping us from having android_log_write_list() clear the data that it has written out from the context, effectively performing this reset automatically?

Otherwise this is fine, just wondering the above.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>