Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Resetting it was not part of the API because we still want(?) to support resending the content on -EBUSY.
> 
> I've been wondering why we let that situation happen. What would go wrong if we just blocked?

IMS: all in the name of a compromise to please all.

Blocking would result in priority inversion, logd is a background SCHED_BATCH process (yes, a decision that could be reversed, but was based on power consumption pushback).

Blocking _would_ introduce heisenberg uncertainty (adding debugging changes behavior), traded off for logging uncertainty (lost messages).

Blocking consumes power. Allowing logs to be lost, especially with the prevalence of DOS attacks on logging we see on a weekly or monthly basis during development, did result in reduced power. I introduced blocking for but a moment in the history in 2014 and had to remove it _quickly_, so understandably bruised ;-}

All logging AFAIK, on all platforms (including inside the kernel dmesg) is subject to loss of content in the name of never blocking or introducing interference. There are reasons for using unconnected frames to propagate system logs on big iron ...

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>