adb: add multiple argument handling for services.
Added the ability to pass more than one argument from the caller
to the service.  Cleaned up all service parsing to drop length
constants and instead strlen(<command>).  Noticed that thread
name for disabling verity was verity-on and enabling verity was
verity-off, corrected the thread names.

SideEffects: If a second argument is passed to a command, it
             may break.
Test: manual
Bug: 109821005
Change-Id: I88d4be65337ff899e4f815e8b1b30288973e0851