Update patch set 10
Patch Set 10: Verified+1

No change due to the PS7 request advised. The deletion and recreation decision in adb remount and adb disable-verity needs to be made _in_ setup function in fs_mgr, if an alternate directory is specified, then it should tear down and rebuild anew as needed, even _possibly_ copying content in doing so. Since we dropped multiple backing directories, in this phase we only have _one_ directory, and no means for us to pass it in, that clearing work is deferred for the patches that add these other features.

Patch-set: 10
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1