Be permissive about badly formed elf files.
Allow the program header and/or section headers to be slightly
malformed. There is no real reason to require that all of the program
headers or section headers to be present. Everything has sane defaults,
so any missing information won't cause any problems.

This gets rid of the warning that occurs any time an elf is
loaded from memory. In memory elf files never contain all of the headers.

Update tests to test these new cases.

Test: Builds and unit tests all pass.
Change-Id: Iaefe2cd6b6c965a01ed425a112d6afae339f3b78