Update patch set 1
Patch Set 1:

I'm not sure, there are a number of side effects that ripple through the Android system when these properties are changed (job scheduling looks for battery charger connected and healthy battery for various cases, etc.), and it can be a bit scary to change them.  Plus, we've re-done things for the P release, the documentation for which is going to Early Access Program now.  If there's a need to get OC MR1 fixed separately i can look more closely at this, maybe drag in somebody who's been working on this more recently.

Patch-set: 1
Reviewer: Gerrit User 1004277 <1004277@85c56323-6fa9-3386-8a01-6480fb634889>