Update patch set 5
Patch Set 5:

(1 comment)

> Patch Set 3:
> 
> (2 comments)
> 
> Is the idea to verify that logical partitions exist before we flash them? If so, would we ever want to create them on-demand once we have fastbootd support in place?

Actually, the CL is to make sure that super and partitions that are part of super(all of which reside in $OUT) do not get flashed in a flashall command unless the device's partition table supports them. You are right that we will need to add support to create logical partitions on-demand once fastbootd support is ready.

Patch-set: 5
Reviewer: Gerrit User 1120612 <1120612@85c56323-6fa9-3386-8a01-6480fb634889>