fastbootd: Support flashall with logical partitions.
Currently, fastboot flashall will try to flash every partition in the
image list. For devices with logical partitions, this will cause those
partitions to be flashed twice (once via super.img, and again for each
individual image).

This patch skip flashing super.img during "flashall" if the device is
using fastbootd.

In addition, we also resize logical partitions before flashing them.
This occurs in two steps: an initial step to resize them to 0 (which
removes their extents), and then a follow-up step to size them to the
on-disk image size. This ensures more optimal extent allocation.

Bug: 78793464
Test: fastboot flashall can flash vendor, product, product_servers, and
      does not flash super.

Change-Id: I18c2291e0d0cb10de729c5646b75ac84dde8c871