Update patch set 4
Patch Set 4:

(5 comments)

> Patch Set 2:
> 
> > Patch Set 2: Code-Review+1
> > 
> > (1 comment)
> > 
> > Ok for getting a first version in, but should we wait for system/apexd to be created? I can anyway work on the binder changes for init independently.
> 
> We can merge here first and move it to other place when the git is created. I've requested the creation of /system/apex (not apexd) to host apexd, packager, and other libs/tools that we might need in the future.

Elliot is the only owner of /system/core. Rather than bothering him with the review, we can finish the review of this CL here and you can patch it in if you need it, and I'll port it to /system/apex when the repo is created.

Patch-set: 4
Reviewer: Gerrit User 1305591 <1305591@85c56323-6fa9-3386-8a01-6480fb634889>