Update patch set 8
Patch Set 8:

FYI: I'll be OOO until next Wednesday, so these CLs won't be submitted till after then.  The build side of the CL still needs work.

> > This should still works, ramdisk device should specify "/system" mount point in fstab-dt.
> > https://android.googlesource.com/platform/system/libvintf/+/91c1b2703bf422786f3d698fb60da5a33fcd83a2/VintfObjectRecovery.cpp#56
> > 
> > But this should change to always return mSystemFileSystem ("/mnt/system") unconditionally.
> > https://android.googlesource.com/platform/system/libvintf/+/91c1b2703bf422786f3d698fb60da5a33fcd83a2/VintfObjectRecovery.cpp#99
> > 
> > Because system.img layout is unified to system-as-root, no matter it's ramdisk or no-ramdisk.
> > Ramdisk devices will have SwitchRoot("/system").
> > 
> 
> Filed b/112586588 to track the change to vintf.

I'll make this change when these changes are submitted.  Good catch.

> 
> > Btw, in P and before, "ro.build.system_root_image" means no-ramdisk and system-as-root (system.img should be mounted as "/", as opposed to "/system").
> > 
> > I'm not sure should we introduce another property for no-ramdisk/ramdisk and always set ro.build.system_root_image=True in Q.
> > 
> > 
> > P and before
> >    ramdisk  |  system-as-root
> > ----------------------------------
> >    O        |          X
> >    X        |          O
> > 
> > 
> > Q and after
> >     ramdisk  |  system-as-root
> > ----------------------------------
> >     O        |          X      [1]
> >     O        |          O
> >     X        |          O
> > 
> > [1] Probably no such config in Q and going forward? 
> >     Legacy ramdisk devices upgraded from O -> P -> Q will have SwitchRoot("/system"), where system.img should be mounted under "/".
> 
> Don't think we'll have config [1]. And it seems the current ro.build.system_root_image will be always true (or could be removed after we clean up all the users).

Correct system.img will always be compiled as if it were system-as-root.  

ro.build.system_root_image will indicate whether or not system.img is used to boot the system or if the ramdisk is, since there are still differences that we need to account for between those two scenarios.  It seems that a majority of users of ro.build.system_root_image are actually used for that distinction than not.  The vintf one is the only one that I see that will go away.

Patch-set: 8
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0