Update patch set 20
Patch Set 20:

> Patch Set 20:
> 
> Actually, a question: How does this interact with https://android-review.googlesource.com/c/platform/system/core/+/705933 ?
> 
> That change talks about having a 'true' vs 'auto'.  Shouldn't setting that to true be enough to override this check?  
> 
> In general, we should keep the number of developer knobs to a minimum here (well, everywhere...).  If we already plan to give developers an option to always enable overlayfs, then there's no reason to give them a way to override the max free percentage.

I believe that c/705933 is in dispute and is being allowed to be stagnant for the moment because it is so far down the stack and only makes sense to address when there is more than one possible backing store for overlayfs. Since /data is the only (public) suggestion and I am having troubles 'selling' it, then c/705933 will not happen.

c/705933 is actually there to _disable_ overlayfs when the automation is ill advised, and more importantly picks the _wrong_ backing storage automatically allowing us to specify exactly which one must be; The fact it can also force enable (just as this one can also force-enable) is only a current side-effect and can be eliminated from the option. It was added for 'completeness' only.

Other alternatives for backing storage could be mutually exclusive (eg: /cache for non-A/B and /mnt/scratch for A/B; never specifying /data in the list) also making c/705933 largely unnecessary, or optional.

This CL originally had _no_ tunables, 1% just makes sense _forever_, I only ended added the tunables because I saw a testing opportunity and have used it throughout development. If you want, I would be _pleased_ to revert this to the original form, or perhaps split out ro.adb.remount.* and leave _this_ CL to a hard coded value of 1%.

Come to think of it, splitting makes sense ...

Patch-set: 20
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0