Update patch set 1
Patch Set 1:

Bug? I kind of like a property not to be empty. If it is supposed to be an int, it shouldn't ever be set to ''. If this is for more complicated use cases, those should be expressed in a more structured IPC language.

Patch-set: 1
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>