adb: win32: use Windows security mitigation features: ASLR, DEP
The --dynamicbase linker flag enables use of Address Space Layout
Randomization (ASLR) on Windows Vista and newer. It causes the code
segment / module to be loaded at unpredictable addresses to make
exploits harder to construct.

The --nxcompat linker flag enables use of Data Execution Protection
(DEP) on Windows XP SP2 and newer. It causes the process to be killed if
the process tries to execute code on memory pages not marked executable.
This makes exploits harder to construct because the stack is then not
executable, other data pages aren't, etc.

These linker flags set bits in the executable PE header and can be
verified with dumpbin /headers, Task Manager, or Process Explorer.

These bits are already set on AdbWinApi.dll and AdbWinUsbApi.dll.

Test: Used Process Explorer to check the load address before and after
rebooting on Windows 10.
Test: dumpbin /headers adb.exe
Test: Process Explorer: checked the DEP and ASLR columns.
Test: Task Manager: checked the DEP column.

Bug: https://issuetracker.google.com/36952278
Change-Id: I0638518ede5e3d06fce4a482001abd20f47224cb
Signed-off-by: Spencer Low <CompareAndSwap@gmail.com>