Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> why aren't we disabling mac and windows?

The issue is not in fuzzy_fastboot, the issue is the Android.bp file for libfastboot2 is missing the required libraries includes for each platform. 

Since eventually, libfastboot (used by the fastboot tool) will be killed and replaced with libfastboot2, it is important libfastboot2 also builds properly on all platforms.

Patch-set: 3
Reviewer: Gerrit User 1336956 <1336956@85c56323-6fa9-3386-8a01-6480fb634889>