Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > I am building one last time on mac, but this patch should fix everything.
> > 
> > The fix actually ended up being two things. Firstly the missing "usb_*.cpp"files in libfastboot2's Android.bp file. 
> > 
> > The second issue is the build system when building something (fuzzy_fastboot) that uses a static library (libfastboot2) that includes shared libraries (all the usb crap), does not seem to transitively include these shared library dependencies (all the usb crap) when linking. You must manually include these transitive shared library dependencies.
> 
> +dwillemsen, because although this was always broken with make, i thought soong was supposed to fix it?

Nope, Soong would allow us to fix that, but it's a decently large behavioral change that would need a lot of verification, and we didn't want the behaviors of make and Soong to diverge, as that makes it harder to convert. So we're planning on doing this once we're closer to a full conversion, and have written some tools to verify the effective changes. 

We will reorder static library dependencies based on transitive dependencies.

Patch-set: 4
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>