fs_mgr: overlayfs create scratch using a Dynamic Logical Partition
Now that we support /mnt/scratch (kScratchMountPoint) passively, if it
does not exist and support for Dynamically Resized Partitions exists
on the device, then add support for setup and teardown of the
associated metatdata for the backing storage.

Added fs_mgr_overlayfs_scratch_present() as hint to init to make sure
the device gets created.

Test: manual
Bug: 109821005
Change-Id: I5e3f66778ea9e937c839e25541fedb6fbf34ab50