fs_mgr: create internal fs_mgr_system_root_image()
Switch checking for "ro.build.system_root_image" because in all cases
if there is a "/system" specified in the fstab, then it is false,
otherwise true.  Place logic into fs_mgr_system_root_image().

If this determination is wrong, which was the case during first stage
init because property service not yet initialized, you will see the
following log message:

    [libfs_mgr]Could not find verity device for mount point: /: No such device or address

Test: manual
Bug: 109821005
Change-Id: I6bbfc0af63e1526e52a1c6365241dbc1d5ee7c6a