Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 1: Code-Review-1
> > 
> > This is a busy loop. Can't we poll(2) or use exponential backoff or something?
> 
> We probably can't use poll, since libbase gets used on windows as well. (There's WSAPoll, but it only works on SOCKETs). select has the problem of being limited FD numbers below an arbitrary limit.
> 
> Sleeping might be the best option? ¯\_(ツ)_/¯

+1 on "Where did this actually crop up in practice?" The answer to this question may inform the right technical approach.

In the meantime, I assume we want to be generic. If that's the case, we can use WaitForSingleObject on Windows, can't we? We can get the handle with __get_osfhandle. (Yes, in practice, WFSO works on sockets, although if you want to WSA-ly correct, you can detect a socket and use the corresponding function.)

Patch-set: 2
Reviewer: Gerrit User 1130394 <1130394@85c56323-6fa9-3386-8a01-6480fb634889>