Update patch set 2
Patch Set 2:

This case is needed if we ever plan to support non-blocking reads. In my case, it was requested that I use this routine to replace a read I was doing on a non-blocking fd created using pipe.

I'm happy to abandon this and use my own custom code in that one case if we don't really want to support this particular case.

Patch-set: 2
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>