Revert "Remove the "needs_erase" workaround."
This reverts commit 59efd440ebb251793d13dd5ea52c1fc1e913ba6d.

In Android verified boot design, the metadata is at the end of the partition. It could be possible to flash an image without verity metadata (e.g., GSI). Without erase, the residual verity metadata for previous image will be used to verify the newly flashed image. This will leads to verity error.

Bug: 112876292
Change-Id: Iaeb0e6d89af1c6ea8424c362d09377d938e2242c
Test: `fastboot -w flashall`