Update patch set 1
Patch Set 1:

This is probably my mistake for not catching this in the reviews, but if FastBootDriver actually owns Transport, then its transport_ member really should be a unique_ptr.  Transport::Close() should happen automatically when its destructed too, to remove the need for an explicit destructor of FastBootDriver.

Alternatively, maybe we should go back to the case where FastBootDriver doesn't own transport_?  For this specific case it makes sense to have the ownership outside, since Transport is an interface and there's a specialized concrete type that fuzzy fastboot wants to interact with.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>