ueventd+libdm: populate /dev/mapper link
Since DM_NAME= is not sent or interpreted with ueventd message,
instead probe /sys/devices/virtual/block/dm-X/dm/name when
instantiating.  Since it is too late when the dm is deleted,
perform the symbolic link delete inline within libdm in
DeleteDevice.  The uevent to follow deletes the master node.

Test: manual
Bug: 109821005
Change-Id: I6d9e467970dfdad7b67754ad61084964251eb05f