ueventd: populate /dev/mapper link
Since DM_NAME= is not sent (bug?) or interpreted with ueventd
message, instead probe /sys/devices/virtual/block/dm-X/dm/name when
instantiating.  Cache the value for later delete.

NB: For Android, the deletion of the nodes will only happen in the
context of fastbootd; otherwise the links created can be considered
persistent.

Test: manual
Bug: 109821005
Change-Id: I6d9e467970dfdad7b67754ad61084964251eb05f