healthd: charger uses shared libraries.
With the recent change in [1], /sbin/charger always gets included into
system.img (as opposed to ramdisk) and started by second-stage-init. It
no longer needs to be statically linked. This CL switches it to use
shared libraries, and moves it to /system/bin/charger.

With this change, charger no longer gets installed into recovery image,
which will be handled by follow-up CL.

[1] https://android-review.googlesource.com/c/platform/build/+/728287

Bug: 73660730
Test: Boot into charger mode on walleye. Check that charger UI works.
Test: Run charger_test on walleye.
Change-Id: Idff70cfad09c825cdba4d35a471ff7ecd5f8a8f7