Update patch set 2
Patch Set 2:

Sounds good.  I was figuring it was a set up issue too, since this change should be a no-op unless androidboot.force_normal_boot is set or /system_recovery_mount is in the fstab, neither of which should be the case by default.

tbao@, enh@ any comments?  Shouldn't hurt to get this in now even if we don't use it and/or revert it later.

Also +howardsoc@ as an fyi too.

Patch-set: 2
Reviewer: Gerrit User 1135739 <1135739@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>