Update patch set 4
Patch Set 4: Verified+1 Presubmit-Ready+1

(2 comments)

I have changed the code that depends on this to _find_ the name of the dm node and use it rather than requiring /dev/block/mapper/<logical-partition> so the only reasons for this adjustment is to add the API (/dev/block/mapper) to aid either debugging or partners, nothing _depends_ on this today.

If the API is viewed as unneeded complication, or one we do not wish to support, we can abandon this (for posterity).

Patch-set: 4
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1
Label: Presubmit-Ready=+1