Update patch set 2
Patch Set 2:

(3 comments)

Why is this not C++?  Memory consumption is not a valid reason against that.  This would be 100% more readable if it were C++ and used libbase for all of the functionality that that provides.  Not to mention safer than the fixed length char arrays used here.

Patch-set: 2
Reviewer: Gerrit User 1054468 <1054468@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>