Update patch set 1
Patch Set 1:

Ack; as mentioned in person, we really should work with the security team on how to do this before the SEPolicy is loaded, since otherwise overlayfs can't be used for updating the SEPolicy and that's an important use case.

Probably no need to have the second insurance mount.  Either it works the first time or we should understand why it failed.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>