Update patch set 1
Patch Set 1:

Thanks for fixing that! With that change LOG_ALWAYS_FATAL_IF (and thus CHECK) should be usable in constexpr functions - and when used in constexpr context (e.g. in static_assert or in template argument) it would detect an error in compilation time - exactly as expected.

Patch-set: 1
Reviewer: Gerrit User 1069991 <1069991@85c56323-6fa9-3386-8a01-6480fb634889>