init: move InitKernelLogging() to first stage init
InitKernelLogging() closes the kernel provided stdin/stdout/stderr,
and therefore should be done as soon as possible.  If these are left
open, it's possible to trigger SELinux audits when transitioning
domains to the init domain if these fds point to a file that the init
domain does not have permission to read.

Bug: 117281017
Test: no audits when booting marlin.
Change-Id: If27edab5c32b27765e24c32fbed506ef625889de