Move root mountpoint and symlink rules to image generation
The root directory requires a set of mountpoints and symlinks that
were previously created using a LOCAL_POST_INSTALL_CMD on the
init and init.environ.rc modules.  If any of these files were removed
there was no way for the build system to know to recreate them.
Move them to the rules that build the ramdisk, system-as-root and
recovery images instead.

Bug: 116661769
Test: m checkbuild
Change-Id: I5d2b81ff7a1c7844b2b7783143a49dd4dc22526d