logd: use unique_ptr for LastLogTimes
There are a few locations where either a LogTimeEntry instance is
removed from LastLogTimes without being deleted (and its associated
thread still running) or where a LogTimeEntry instance is deleted
without being removed from LastLogTimes.

This change uses unique_ptr in LastLogTimes to ensure that the above
two cases are impossible as a LogTimeEntry instance is only deleted
when it is erased from the list.

Test: not yet
Change-Id: Ibe8651e7d530c5e9a8d6ce3150cd247982887cbe