Update patch set 1
Patch Set 1: Code-Review-1

I saw this and the internal CL and the article and I'm not quite sure I like any of these answers.

The article is talking about comparing to a literal 0xFF, which is an int and therefore the char is promoted to a signed int and the comparison fails, but that's not the case in this original code.  It's comparing a char with a char, so there should be no promotion and the comparison should succeed.  Are we sure there's not a latent compiler bug here that is doing this promotion unnecessarily here?

I don't like this CL taking an int as the parameter either, since that isn't really the right signature here.  It should take a 'char' since values larger than a char are meaningless.

Does the internal CL with '\xFF' solve this issue?  I'd rather that be submitted than changing the function definition and an investigation opened for why this comparison is failing (since it's not like the one in the article).

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1